requirements.txt
selenium==4.23.1
webdriver_manager==4.0.2
ffmpeg-python==0.2.0